move logic from page to scene
