+++ src/pages/aoi/component.jsx
@@ -6,24 +6,10 @@ import HalfEarthLogo from 'components/half-earth-logo';
 import styles from './styles.module.scss';
 import uiStyles from 'styles/ui.module.scss';
 
-const AreaOfInterestPageComponent = ({
-  aoiData,
-  geometry,
-  userConfig,
-  activeLayers,
-  sceneSettings,
-}) => (
+const AreaOfInterestPageComponent = () => (
   <>
     <HalfEarthLogo className={cx(styles.hideOnPrint,uiStyles.halfEarthLogoTopLeft)}/>
-    {geometry &&
-      <AreaOfInterestScene
-        aoiData={aoiData}
-        geometry={geometry}
-        userConfig={userConfig}
-        activeLayers={activeLayers}
-        sceneSettings={sceneSettings}
-      />
-    }
+    <AreaOfInterestScene />
   </>
 );
