+++ src/scenes/aoi-scene/selectors.js
@@ -1,0 +1,30 @@
+import { createSelector, createStructuredSelector } from 'reselect';
+import { selectGlobeUrlState } from 'selectors/location-selectors';
+import { selectAoiId, getAoiGeometry } from 'selectors/aoi-selectors';
+
+import aoiSceneConfig from './config';
+
+const selectUserConfig = ({ userConfig }) => userConfig || null;
+const selectPrecalculatedLayerSlug = ({location}) => location.query && location.query.precalculatedLayer;
+
+const getGlobeSettings = createSelector([selectGlobeUrlState],
+  (globeUrlState) => {
+  return {
+    ...aoiSceneConfig,
+    ...globeUrlState
+  }
+})
+
+export const getActiveLayers = createSelector(getGlobeSettings, globeSettings => globeSettings.activeLayers);
+const getGlobeUpdating = createSelector(getGlobeSettings, globeSettings => globeSettings.isGlobeUpdating);
+
+
+export default createStructuredSelector({
+  aoiId: selectAoiId,
+  userConfig: selectUserConfig,
+  activeLayers: getActiveLayers,
+  sceneSettings: getGlobeSettings,
+  isGlobeUpdating: getGlobeUpdating,
+  aoiStoredGeometry: getAoiGeometry,
+  precalculatedLayerSlug: selectPrecalculatedLayerSlug,
+});
